moved tests to input test file
